% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpf-package.R
\docType{package}
\name{cpf-package}
\alias{cpf}
\alias{cpf-package}
\title{cpf: Tools for handling cpf data}
\description{
This package provides tools for handling cpf data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arturhgq/cpf}
  \item Report bugs at \url{https://github.com/arturhgq/cpf/issues}
}

}
\author{
\strong{Maintainer}: Artur Quirino \email{contact.last@arturhgq.com} (\href{https://orcid.org/0000-0001-8071-6348}{ORCID})

}
\keyword{internal}
