% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpf.R
\name{fix_cpf}
\alias{fix_cpf}
\title{Make cpf strings consistent across different patterns}
\usage{
fix_cpf(.cpf)
}
\arguments{
\item{.cpf}{CPF string}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function makes cpf strings consistent across different patterns
}
\examples{
  fake_cpf <- c(
   "111,232-123.23",
   "11122233323",
   "234..222--454,,00"
 )
 fix_cpf(fake_cpf)
}
