% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valida_cpf.R
\name{valida_cpf}
\alias{valida_cpf}
\title{Valida CPF}
\usage{
valida_cpf(x = "52998224725")
}
\arguments{
\item{x}{String com o numero do CPF com 11 digitos}
}
\description{
Retorna "TRUE" para um numero de CPF valido
}
\examples{
valida_cpf("52998224725")
}
