% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpf.R
\name{validate_cpf}
\alias{validate_cpf}
\title{Validate cpf}
\usage{
validate_cpf(.data, .cpf, .fix_cpf = TRUE, .clear_steps = TRUE)
}
\arguments{
\item{.data}{data frame}

\item{.cpf}{cpf variable}

\item{.fix_cpf}{\code{\link[=fix_cpf]{fix_cpf()}}}

\item{.clear_steps}{delete the variables used to validate cpfs}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function validates cpfs by calculating of the check digits
}
\examples{

data.frame(
 cpf = c(
   "22233344",
   "423.643.534-29",
   "458.613.523-23",
   13434563434,
   NA
   )
 ) -> data
 validate_cpf(data, cpf)
}
